migrate WidgetFactory to TypeScript and changed the build process to rollup
