+++ src/main/js/bundles/dn_lineofsight/manifest.json
@@ -85,7 +85,6 @@
     "components": [
         {
             "name": "Config",
-            "impl": "./LineOfSightWidgetFactory",
             "provides": [
                 "dijit.Widget",
                 "dn_lineofsight.Widget"
